+++ packages/exception/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rym-lib/exception
 
+## 1.2.6
+
+### Patch Changes
+
+- feat(query-module): add function-based rules support #50
+
 ## 1.2.5
 
 ### Patch Changes