+++ packages/query-module/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rym-lib/query-module
 
+## 1.2.6
+
+### Patch Changes
+
+- feat(query-module): add function-based rules support #50
+- Updated dependencies
+  - @rym-lib/exception@1.2.6
+
 ## 1.2.5
 
 ### Patch Changes